pyproject.toml
[project]
name = "speech2text"
version = "0.1.0"
description = "Audio Transcription Desktop Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "pyside6>=6.10.1",
    "python-dotenv>=1.2.1",
    "pyinstaller>=6.11.1",
]
